requirements.txt
beautifulsoup4==4.5.1
bs4==0.0.1
dnspython==1.14.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.5.3
requests==2.11.1
six==1.10.0
sleekxmpp==1.3.1
